requirements.txt
streamlit
trimesh
numpy
pillow
scikit-learn
scipy